refactor(config,context): shrink ai_provider function and split dependency-parser/project-shape modules
