+++ src/quodeq/config/_dependency_parsers.py
@@ -11,308 +11,32 @@ raw text, which produced false positives in two classes:
 
 For files we don't know how to parse, ``DisciplineRegistry`` falls back to the
 existing substring behaviour — see ``_discipline_detection.py``.
+
+Python (pyproject.toml, requirements.txt) matchers live in
+``_dependency_parsers_python.py``; package.json/Cargo.toml/go.mod/
+composer.json/pom.xml/Gradle matchers live in
+``_dependency_parsers_compiled.py`` — both split out to keep this module
+under the size ratchet's 300-line cap. All ``has_*`` names stay re-exported
+from here.
 """
 from __future__ import annotations
 
-import json
 import re
 import tomllib
-import xml.etree.ElementTree as ET
 from typing import Callable
 
-# PEP 503: package names are case-insensitive and ``[-_.]+`` normalize to ``-``.
-_PEP503_SEP = re.compile(r"[-_.]+")
-
-
-def _normalize_pep503(name: str) -> str:
-    return _PEP503_SEP.sub("-", name).strip().lower()
-
-
-def _parse_pep508_name(spec: str) -> str:
-    """Extract the package name from a PEP 508 requirement string.
-
-    Handles version specifiers, extras, environment markers, and direct URLs:
-    ``django>=4`` → ``django``; ``uvicorn[standard]==0.30`` → ``uvicorn``;
-    ``foo @ git+https://...`` → ``foo``; ``bar; python_version<'3.10'`` → ``bar``.
-    """
-    name = spec.strip()
-    if ";" in name:
-        name = name.split(";", 1)[0].strip()
-    if "[" in name:
-        name = name.split("[", 1)[0].strip()
-    # Order matters: longer operators before shorter (=== before ==, >= before >).
-    for op in ("===", "==", ">=", "<=", "~=", "!=", ">", "<", "@", " ", "\t"):
-        if op in name:
-            name = name.split(op, 1)[0].strip()
-    return _normalize_pep503(name)
-
-
-def _names_from_list(items: object) -> set[str]:
-    if not isinstance(items, list):
-        return set()
-    return {_parse_pep508_name(s) for s in items if isinstance(s, str) and s.strip()}
-
-
-def _names_from_dict_keys(items: object) -> set[str]:
-    if not isinstance(items, dict):
-        return set()
-    return {_normalize_pep503(k) for k in items if isinstance(k, str)}
-
-
-# --- pyproject.toml ----------------------------------------------------------
-
-
-def _pyproject_dep_names(content: str) -> set[str]:
-    try:
-        data = tomllib.loads(content)
-    except tomllib.TOMLDecodeError:
-        return set()
-
-    names: set[str] = set()
-    project = data.get("project")
-    if isinstance(project, dict):
-        names |= _names_from_list(project.get("dependencies"))
-        optional = project.get("optional-dependencies")
-        if isinstance(optional, dict):
-            for v in optional.values():
-                names |= _names_from_list(v)
-
-    poetry = data.get("tool", {}).get("poetry") if isinstance(data.get("tool"), dict) else None
-    if isinstance(poetry, dict):
-        names |= _names_from_dict_keys(poetry.get("dependencies"))
-        names |= _names_from_dict_keys(poetry.get("dev-dependencies"))
-        groups = poetry.get("group")
-        if isinstance(groups, dict):
-            for g in groups.values():
-                if isinstance(g, dict):
-                    names |= _names_from_dict_keys(g.get("dependencies"))
-
-    # PEP 735 dependency groups (used by uv and others).
-    dep_groups = data.get("dependency-groups")
-    if isinstance(dep_groups, dict):
-        for v in dep_groups.values():
-            names |= _names_from_list(v)
-
-    return names
-
-
-def has_pyproject_dependency(content: str, needle: str) -> bool:
-    return _normalize_pep503(needle) in _pyproject_dep_names(content)
-
-
-# --- requirements.txt --------------------------------------------------------
-
-
-def _requirements_txt_names(content: str) -> set[str]:
-    names: set[str] = set()
-    for raw in content.splitlines():
-        line = raw.split("#", 1)[0].strip()
-        if not line:
-            continue
-        # Skip pip directives (-r, -e, --extra-index-url, etc.).
-        if line.startswith("-"):
-            continue
-        name = _parse_pep508_name(line)
-        if name:
-            names.add(name)
-    return names
-
-
-def has_requirements_txt_dependency(content: str, needle: str) -> bool:
-    return _normalize_pep503(needle) in _requirements_txt_names(content)
-
-
-# --- package.json ------------------------------------------------------------
-
-
-_PACKAGE_JSON_DEP_KEYS = (
-    "dependencies", "devDependencies", "peerDependencies",
-    "optionalDependencies", "bundledDependencies", "bundleDependencies",
+from quodeq.config._dependency_parsers_python import (  # noqa: F401 - re-export
+    has_pyproject_dependency,
+    has_requirements_txt_dependency,
 )
-
-
-def _package_json_names(content: str) -> set[str]:
-    try:
-        data = json.loads(content)
-    except (json.JSONDecodeError, ValueError):
-        return set()
-    if not isinstance(data, dict):
-        return set()
-    names: set[str] = set()
-    for key in _PACKAGE_JSON_DEP_KEYS:
-        v = data.get(key)
-        if isinstance(v, dict):
-            names.update(k.lower() for k in v if isinstance(k, str))
-        elif isinstance(v, list):  # bundledDependencies is a list of strings
-            names.update(s.lower() for s in v if isinstance(s, str))
-    return names
-
-
-def has_package_json_dependency(content: str, needle: str) -> bool:
-    return needle.strip().lower() in _package_json_names(content)
-
-
-# --- Cargo.toml --------------------------------------------------------------
-
-
-def _cargo_dep_names(content: str) -> set[str]:
-    try:
-        data = tomllib.loads(content)
-    except tomllib.TOMLDecodeError:
-        return set()
-    names: set[str] = set()
-    for key in ("dependencies", "dev-dependencies", "build-dependencies"):
-        v = data.get(key)
-        if isinstance(v, dict):
-            names.update(k.lower() for k in v if isinstance(k, str))
-    # Workspace dependencies: [workspace.dependencies]
-    workspace = data.get("workspace")
-    if isinstance(workspace, dict):
-        wdeps = workspace.get("dependencies")
-        if isinstance(wdeps, dict):
-            names.update(k.lower() for k in wdeps if isinstance(k, str))
-    # Target-specific dependencies: [target."cfg(...)".dependencies]
-    target = data.get("target")
-    if isinstance(target, dict):
-        for tcfg in target.values():
-            if not isinstance(tcfg, dict):
-                continue
-            for key in ("dependencies", "dev-dependencies", "build-dependencies"):
-                v = tcfg.get(key)
-                if isinstance(v, dict):
-                    names.update(k.lower() for k in v if isinstance(k, str))
-    return names
-
-
-def has_cargo_dependency(content: str, needle: str) -> bool:
-    return needle.strip().lower() in _cargo_dep_names(content)
-
-
-# --- go.mod ------------------------------------------------------------------
-
-
-def _go_mod_modules(content: str) -> set[str]:
-    """Return the set of module paths declared in ``require`` directives."""
-    modules: set[str] = set()
-    in_block = False
-    for raw in content.splitlines():
-        line = raw.strip()
-        if not line or line.startswith("//"):
-            continue
-        if line.startswith("require ("):
-            in_block = True
-            continue
-        if in_block and line.startswith(")"):
-            in_block = False
-            continue
-        if line.startswith("require "):
-            line = line[len("require "):].strip()
-        elif not in_block:
-            continue
-        # Strip inline comments and parse the first whitespace-delimited token.
-        line = line.split("//", 1)[0].strip()
-        parts = line.split(None, 1)
-        if parts:
-            modules.add(parts[0])
-    return modules
-
-
-def has_go_mod_module(content: str, needle: str) -> bool:
-    """Match a module path. ``foo/bar`` matches both ``foo/bar`` and ``foo/bar/v2``."""
-    needle = needle.strip()
-    if not needle:
-        return True
-    prefix = needle + "/"
-    for mod in _go_mod_modules(content):
-        if mod == needle or mod.startswith(prefix):
-            return True
-    return False
-
-
-# --- composer.json -----------------------------------------------------------
-
-
-def _composer_dep_names(content: str) -> set[str]:
-    try:
-        data = json.loads(content)
-    except (json.JSONDecodeError, ValueError):
-        return set()
-    if not isinstance(data, dict):
-        return set()
-    names: set[str] = set()
-    for key in ("require", "require-dev"):
-        v = data.get(key)
-        if isinstance(v, dict):
-            names.update(k.lower() for k in v if isinstance(k, str))
-    return names
-
-
-def has_composer_dependency(content: str, needle: str) -> bool:
-    return needle.strip().lower() in _composer_dep_names(content)
-
-
-# --- pom.xml (Maven) ---------------------------------------------------------
-
-
-def _pom_coords(content: str) -> set[str]:
-    """Return all groupId / artifactId text values declared in *content*.
-
-    Captures dependencies, parent coords, plugins, and BOM imports — anywhere a
-    ``<groupId>`` or ``<artifactId>`` element appears. Commentary and free text
-    in ``<description>`` / ``<comment>`` are not collected.
-    """
-    try:
-        root = ET.fromstring(content)
-    except ET.ParseError:
-        return set()
-    coords: set[str] = set()
-    for elem in root.iter():
-        # Strip default Maven namespace if present (``{http://maven.apache.org/POM/4.0.0}groupId``).
-        tag = elem.tag.rsplit("}", 1)[-1]
-        if tag in ("groupId", "artifactId") and elem.text:
-            coords.add(elem.text.strip())
-    return coords
-
-
-def has_pom_xml_dependency(content: str, needle: str) -> bool:
-    """Substring-match *needle* against extracted Maven coordinates.
-
-    Substring (not exact) because rule needles like ``spring-boot`` must match
-    artifactIds like ``spring-boot-starter-web``. Description / comment text is
-    excluded by construction — only ``<groupId>`` / ``<artifactId>`` are
-    considered, so a ``<description>migrating off spring-boot</description>``
-    no longer triggers a false match.
-    """
-    needle_low = needle.strip().lower()
-    if not needle_low:
-        return True
-    return any(needle_low in c.lower() for c in _pom_coords(content))
-
-
-# --- Gradle (Groovy / Kotlin DSL) -------------------------------------------
-
-
-_GRADLE_BLOCK_COMMENT = re.compile(r"/\*.*?\*/", re.DOTALL)
-_GRADLE_LINE_COMMENT = re.compile(r"//[^\n]*")
-
-
-def _strip_gradle_comments(content: str) -> str:
-    return _GRADLE_LINE_COMMENT.sub("", _GRADLE_BLOCK_COMMENT.sub("", content))
-
-
-def has_gradle_dependency(content: str, needle: str) -> bool:
-    """Substring-match against build.gradle / build.gradle.kts with comments stripped.
-
-    Real Gradle parsing requires a JVM; a comment-aware substring is the best
-    we can do without one. Strips ``//`` line and ``/* */`` block comments first
-    so a ``// migrating off spring-boot`` line no longer triggers a match.
-    Coordinates inside string literals (the actual dep declarations) survive.
-    """
-    needle_low = needle.strip().lower()
-    if not needle_low:
-        return True
-    return needle_low in _strip_gradle_comments(content).lower()
-
+from quodeq.config._dependency_parsers_compiled import (  # noqa: F401 - re-export
+    has_cargo_dependency,
+    has_composer_dependency,
+    has_go_mod_module,
+    has_gradle_dependency,
+    has_package_json_dependency,
+    has_pom_xml_dependency,
+)
 
 # --- Gemfile (Ruby DSL) ------------------------------------------------------
