+++ src/quodeq/config/ai_provider.py
@@ -49,35 +49,41 @@ def get_current_provider(
     return get_ai_provider()
 
 
-def _write_env(paths: ConfigPaths, provider: str, api_key_var: str, api_key_value: str) -> None:
+def _build_env_lines(paths: ConfigPaths, provider: str, api_key_var: str, api_key_value: str) -> list[str]:
     lines = [
         "# Quodeq provider config (generated by quodeq configure --ai-cli)",
         f"export AI_PROVIDER={provider}",
     ]
-    if api_key_value:
-        # Reject control characters that would let a malformed input append
-        # extra `export …` lines to the env file.
-        if any(ch in api_key_value for ch in _API_KEY_FORBIDDEN_CHARS):
-            raise ValueError("API key contains forbidden control characters")
-        # SECURITY: The raw API key value is written to the env file in
-        # cleartext because it must be sourced by subprocesses at runtime.
-        # The file is created with mode 0600 (owner read/write only) via
-        # atomic write + fchmod.  This is a known trade-off for local CLI
-        # tools where no platform keychain integration is available.
-        #
-        # For production / shared-server deployments:
-        #   - Use the QUODEQ_API_KEY environment variable directly, or
-        #   - Store the key in a platform keychain (macOS Keychain,
-        #     GNOME Keyring, Windows Credential Manager), or
-        #   - Use a secrets manager (Vault, AWS Secrets Manager, etc.)
-        lines.append(f"export {api_key_var}={api_key_value}")
-        masked = "***" + api_key_value[-4:] if len(api_key_value) > 4 else "****"
-        log_warning(
-            f"API key ({masked}) will be stored in cleartext at "
-            f"{paths.env_file} (mode 0600). For production deployments, "
-            f"prefer the QUODEQ_API_KEY environment variable, a platform "
-            f"keychain, or a secrets manager instead of writing keys to disk."
-        )
+    if not api_key_value:
+        return lines
+    # Reject control characters that would let a malformed input append
+    # extra `export …` lines to the env file.
+    if any(ch in api_key_value for ch in _API_KEY_FORBIDDEN_CHARS):
+        raise ValueError("API key contains forbidden control characters")
+    # SECURITY: The raw API key value is written to the env file in
+    # cleartext because it must be sourced by subprocesses at runtime.
+    # The file is created with mode 0600 (owner read/write only) via
+    # atomic write + fchmod.  This is a known trade-off for local CLI
+    # tools where no platform keychain integration is available.
+    #
+    # For production / shared-server deployments:
+    #   - Use the QUODEQ_API_KEY environment variable directly, or
+    #   - Store the key in a platform keychain (macOS Keychain,
+    #     GNOME Keyring, Windows Credential Manager), or
+    #   - Use a secrets manager (Vault, AWS Secrets Manager, etc.)
+    lines.append(f"export {api_key_var}={api_key_value}")
+    masked = "***" + api_key_value[-4:] if len(api_key_value) > 4 else "****"
+    log_warning(
+        f"API key ({masked}) will be stored in cleartext at "
+        f"{paths.env_file} (mode 0600). For production deployments, "
+        f"prefer the QUODEQ_API_KEY environment variable, a platform "
+        f"keychain, or a secrets manager instead of writing keys to disk."
+    )
+    return lines
+
+
+def _write_env(paths: ConfigPaths, provider: str, api_key_var: str, api_key_value: str) -> None:
+    lines = _build_env_lines(paths, provider, api_key_var, api_key_value)
     fd, tmp_path = tempfile.mkstemp(dir=str(paths.env_file.parent), suffix=".tmp")
     closed = False
     try:
